+++ src/main/java/hellfirepvp/astralsorcery/client/gui/journal/page/JournalPageAttunementRecipe.java
@@ -66,9 +66,9 @@ public class JournalPageAttunementRecipe implements IJournalPage {
         private void renderAltarSlot(float offsetX, float offsetY, float zLevel, java.util.List<ItemStack> stacks) {
             if(stacks == null || stacks.isEmpty()) return;
 
-            int select = ((ClientScheduler.getClientTick() + ((int) offsetX) * 40 + ((int) offsetY) * 40) / 20);
+            long select = ((ClientScheduler.getClientTick() + ((int) offsetX) * 40 + ((int) offsetY) * 40) / 20);
             select %= stacks.size();
-            ItemStack draw = stacks.get(select);
+            ItemStack draw = stacks.get((int) select);
 
             TextureHelper.refreshTextureBindState();
             GL11.glPushMatrix();
