+++ src/main/java/hellfirepvp/astralsorcery/client/gui/journal/page/JournalPageRecipe.java
@@ -110,9 +110,9 @@ public class JournalPageRecipe implements IJournalPage {
                 if(expected == null || expected.isEmpty()) expected = recipe.getExpectedStackForRender(srs.rowMultipler, srs.columnMultiplier);
                 if(expected == null || expected.isEmpty()) continue;
 
-                int select = ((ClientScheduler.getClientTick() + srs.rowMultipler * 40 + srs.columnMultiplier * 40) / 20);
+                long select = ((ClientScheduler.getClientTick() + srs.rowMultipler * 40 + srs.columnMultiplier * 40) / 20);
                 select %= expected.size();
-                ItemStack draw = expected.get(select);
+                ItemStack draw = expected.get((int) select);
 
                 GL11.glPushMatrix();
                 GL11.glTranslated(offX + (srs.columnMultiplier * 25), offY + (srs.rowMultipler * 25), zLevel + 60);
