+++ src/main/java/hellfirepvp/astralsorcery/common/crafting/altar/recipes/ConstellationRecipe.java
@@ -152,7 +152,7 @@ public class ConstellationRecipe extends AttunementRecipe {
     }
 
     @Override
-    public void onCraftClientTick(TileAltar altar, int tick, Random rand) {
+    public void onCraftClientTick(TileAltar altar, long tick, Random rand) {
         super.onCraftClientTick(altar, tick, rand);
 
         Vector3 altarVec = new Vector3(altar);
